refactor: rename node ESM helper file to esm.js
