+++ scripts/release/versioning.mjs
@@ -3,7 +3,7 @@ import fg from 'fast-glob';
 import fs from 'node:fs';
 import path from 'node:path';
 import { PACKAGE_ROOT } from './util.cjs';
-import { log, errorHandler, success, getJSON } from '../helpers/index.mjs';
+import { log, errorHandler, success, getJSON } from '../helpers/esm.mjs';
 
 /**
  * Placeholder of element version
